Transform NBS Data to Format Necessary for Exporting to ACNS File

prep_positive() transforms data from NBS snapshot files to the format needed for exports to the ACNS process.

prep_positive(
  .data = load_positive(),
  filter_lab = TRUE,
  labs = c("AEL", "BAPTIST", "CCHS", "POPLAR", "UT"),
  filter_new = FALSE,
  filter_acns = filter_new,
  date = attr(.data, "date")
)

Arguments

.data
Data from NBS snapshot file

filter_lab
Should records from labs reported through the ACNS process be filtered out?

labs
Character vector of labs to filter; the default is all labs reporting directly to SCHD

filter_new
Should only new positives be returned? These are still de-duplicated across the full NBS dataset.

filter_acns
Should duplicates from previous ACNS files be removed?

date
The date attribute for the resulting date_tbl

Value

A date_tbl, which is a tibble with a date attribute